old:
```
Warm-up 1: 0.181 us/op
Warm-up 2: 0.176 us/op
Warm-up 3: 0.182 us/op
Warm-up 4: 0.175 us/op
Warm-up 5: 0.176 us/op
Iteration 1: 0.180 us/op
Iteration 2: 0.179 us/op
Iteration 3: 0.204 us/op
Iteration 4: 0.184 us/op
Iteration 5: 0.203 us/op

  Success: 0.190 ±(99.9%) 0.049 us/op [Average]
  (min, avg, max) = (0.179, 0.190, 0.204), stdev = 0.013
  CI (99.9%): [0.141, 0.239] (assumes normal distribution)


jvm summary:
Benchmark                     Mode  Cnt  Score   Error  Units
IdBenchmark.stringDecodeToId  avgt    5  0.190 ± 0.049  us/op
```

new:
```
Warm-up 1: 0.144 us/op
Warm-up 2: 0.145 us/op
Warm-up 3: 0.143 us/op
Warm-up 4: 0.144 us/op
Warm-up 5: 0.143 us/op
Iteration 1: 0.146 us/op
Iteration 2: 0.144 us/op
Iteration 3: 0.143 us/op
Iteration 4: 0.141 us/op
Iteration 5: 0.143 us/op

  Success: 0.143 ±(99.9%) 0.006 us/op [Average]
  (min, avg, max) = (0.141, 0.143, 0.146), stdev = 0.002
  CI (99.9%): [0.137, 0.150] (assumes normal distribution)


jvm summary:
Benchmark                     Mode  Cnt  Score   Error  Units
IdBenchmark.stringDecodeToId  avgt    5  0.143 ± 0.006  us/op
```